--- a/sample/FasterKvCache.Sample.ConsoleApp/ObjectFasterKvCache.cs
+++ b/sample/FasterKvCache.Sample.ConsoleApp/ObjectFasterKvCache.cs
@@ -1,4 +1,5 @@
 ﻿using FasterKv.Cache.Core;
+using FasterKv.Cache.Core.Abstractions;
 using FasterKv.Cache.Core.Configurations;
 using FasterKv.Cache.MessagePack;