--- a/src/FasterKv.Cache.Core/Abstractions/IFasterKvCacheSerializer.cs
+++ b/src/FasterKv.Cache.Core/Abstractions/IFasterKvCacheSerializer.cs
@@ -1,8 +1,12 @@
-﻿namespace FasterKv.Cache.Core;
+﻿using System.IO;
+
+namespace FasterKv.Cache.Core;
 
 public interface IFasterKvCacheSerializer
 {
     string Name { get;}
     byte[] Serialize<TValue>(TValue data);
-    TValue? Deserialize<TValue>(byte[] serializerData);
+    TValue? Deserialize<TValue>(byte[] bytes);
+    void Serialize<TValue>(Stream stream, TValue data);
+    TValue? Deserialize<TValue>(byte[] bytes, int length);
 }