--- a/src/FasterKv.Cache.Core/Configurations/ServiceCollectionExtensions.cs
+++ b/src/FasterKv.Cache.Core/Configurations/ServiceCollectionExtensions.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using FasterKv.Cache.Core.Abstractions;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.DependencyInjection.Extensions;
 using Microsoft.Extensions.Logging;