--- a/src/FasterKv.Cache.Core/FasterKv.Cache.Core.csproj
+++ b/src/FasterKv.Cache.Core/FasterKv.Cache.Core.csproj
@@ -6,7 +6,6 @@
     </PropertyGroup>
 
     <ItemGroup>
-      <PackageReference Include="MessagePack" Version="2.4.35" />
       <PackageReference Include="Microsoft.FASTER.Core" Version="2.1.0" />
       <PackageReference Include="System.Text.Json" Version="7.0.0" />
     </ItemGroup>