--- a/src/FasterKv.Cache.Core/FasterKv.Cache.Core.csproj
+++ b/src/FasterKv.Cache.Core/FasterKv.Cache.Core.csproj
@@ -7,7 +7,6 @@
 
     <ItemGroup>
       <PackageReference Include="Microsoft.FASTER.Core" Version="2.1.0" />
-      <PackageReference Include="System.Text.Json" Version="7.0.0" />
     </ItemGroup>
 
 </Project>