--- a/src/FasterKv.Cache.MessagePack/FasterKv.Cache.MessagePack.csproj
+++ b/src/FasterKv.Cache.MessagePack/FasterKv.Cache.MessagePack.csproj
@@ -9,6 +9,7 @@
     </ItemGroup>
 
     <ItemGroup>
+      <PackageReference Include="MessagePack" Version="2.4.59" />
       <PackageReference Include="System.Text.Json" Version="7.0.0" />
     </ItemGroup>