--- a/src/FasterKv.Cache.MessagePack/MessagePackFasterKvCacheSerializer.cs
+++ b/src/FasterKv.Cache.MessagePack/MessagePackFasterKvCacheSerializer.cs
@@ -8,17 +8,7 @@ namespace FasterKv.Cache.MessagePack;
 public sealed class MessagePackFasterKvCacheSerializer : IFasterKvCacheSerializer
 {
     public string Name { get; set; } = "MessagePack";
-
-    public byte[] Serialize<TValue>(TValue data)
-    {
-        return MessagePackSerializer.Serialize(data);
-    }
-
-    public TValue Deserialize<TValue>(byte[] serializerData)
-    {
-        return MessagePackSerializer.Deserialize<TValue>(serializerData);
-    }
-
+    
     public void Serialize<TValue>(Stream stream, TValue data)
     {
         MessagePackSerializer.Serialize(stream, data);