--- a/src/FasterKv.Cache.SystemTextJson/SystemTextJsonFasterKvCacheSerializer.cs
+++ b/src/FasterKv.Cache.SystemTextJson/SystemTextJsonFasterKvCacheSerializer.cs
@@ -1,8 +1,10 @@
-﻿using FasterKv.Cache.Core;
+﻿using System;
+using System.IO;
+using FasterKv.Cache.Core;
 
 namespace FasterKv.Cache.SystemTextJson;
 
-public class SystemTextJsonFasterKvCacheSerializer : IFasterKvCacheSerializer
+public sealed class SystemTextJsonFasterKvCacheSerializer : IFasterKvCacheSerializer
 {
     public string Name { get; set; } = "SystemTextJson";
     public byte[] Serialize<TValue>(TValue data)
@@ -13,5 +15,15 @@ public class SystemTextJsonFasterKvCacheSerializer : IFasterKvCacheSerializer
     public TValue? Deserialize<TValue>(byte[] serializerData)
     {
         return System.Text.Json.JsonSerializer.Deserialize<TValue>(serializerData);
+    }
+
+    public void Serialize<TValue>(Stream stream, TValue data)
+    {
+        System.Text.Json.JsonSerializer.Serialize(stream, data);
+    }
+
+    public TValue? Deserialize<TValue>(byte[] serializerData, int length)
+    {
+        return System.Text.Json.JsonSerializer.Deserialize<TValue>(new Span<byte>(serializerData,0, length));
     }
 }