--- a/src/FasterKv.Cache.SystemTextJson/SystemTextJsonFasterKvCacheSerializer.cs
+++ b/src/FasterKv.Cache.SystemTextJson/SystemTextJsonFasterKvCacheSerializer.cs
@@ -7,15 +7,6 @@ namespace FasterKv.Cache.SystemTextJson;
 public sealed class SystemTextJsonFasterKvCacheSerializer : IFasterKvCacheSerializer
 {
     public string Name { get; set; } = "SystemTextJson";
-    public byte[] Serialize<TValue>(TValue data)
-    {
-        return System.Text.Json.JsonSerializer.SerializeToUtf8Bytes(data);
-    }
-
-    public TValue? Deserialize<TValue>(byte[] serializerData)
-    {
-        return System.Text.Json.JsonSerializer.Deserialize<TValue>(serializerData);
-    }
 
     public void Serialize<TValue>(Stream stream, TValue data)
     {