--- a/tests/FasterKv.Cache.Core.Tests/KvStore/FasterKvStoreObjectTest.cs
+++ b/tests/FasterKv.Cache.Core.Tests/KvStore/FasterKvStoreObjectTest.cs
@@ -1,4 +1,5 @@
-﻿using FasterKv.Cache.Core.Configurations;
+﻿using FasterKv.Cache.Core.Abstractions;
+using FasterKv.Cache.Core.Configurations;
 using FasterKv.Cache.MessagePack;
 using FasterKv.Cache.SystemTextJson;