--- a/tests/FasterKv.Cache.Core.Tests/KvStore/FasterKvStoreTest.Expiry.cs
+++ b/tests/FasterKv.Cache.Core.Tests/KvStore/FasterKvStoreTest.Expiry.cs
@@ -1,4 +1,5 @@
-﻿using FasterKv.Cache.Core.Configurations;
+﻿using FasterKv.Cache.Core.Abstractions;
+using FasterKv.Cache.Core.Configurations;
 using FasterKv.Cache.MessagePack;
 
 namespace FasterKv.Cache.Core.Tests.KvStore;