* Remove MessagePack from FasterKv.Cache.Core dependency。
* Adjusting type modifiers
* Remove unused method

**Before**

| Method |   Type | ThreadCount |       Mean |    Error |   StdDev |    Gen0 | Allocated |
|------- |------- |------------ |-----------:|---------:|---------:|--------:|----------:|
|   Full |  Write |           8 |   332.1 us | 10.30 us |  6.81 us |  4.3945 |  21.24 KB |
|   Full |   Read |           8 |   356.4 us | 12.81 us |  8.47 us | 11.7188 |  55.42 KB |
|   Full | Random |           8 |   365.8 us | 12.20 us |  8.07 us |  8.3008 |  38.36 KB |
|   Full |  Write |           4 |   541.0 us | 10.67 us |  7.06 us |  7.8125 |  40.09 KB |
|   Full |   Read |           4 |   572.1 us | 12.57 us |  6.58 us | 23.4375 | 108.45 KB |
|   Full | Random |           4 |   607.6 us | 13.52 us |  8.04 us | 15.6250 |  74.27 KB |
|   Full |  Write |           1 | 2,078.4 us | 42.35 us | 28.01 us | 31.2500 | 156.76 KB |
|   Full |   Read |           1 | 2,208.0 us | 37.49 us | 24.79 us | 89.8438 |  430.2 KB |
|   Full | Random |           1 | 2,261.6 us | 15.86 us |  9.44 us | 62.5000 | 293.48 KB |

**After**
| Method |   Type | ThreadCount |       Mean |    Error |   StdDev |    Gen0 | Allocated |
|------- |------- |------------ |-----------:|---------:|---------:|--------:|----------:|
|   Full |  Write |           8 |   333.5 us |  6.58 us |  4.35 us |  4.3945 |  21.24 KB |
|   Full |   Read |           8 |   349.1 us | 12.23 us |  8.09 us | 10.7422 |  50.54 KB |
|   Full | Random |           8 |   358.6 us | 11.97 us |  7.92 us |  7.8125 |  35.91 KB |
|   Full |  Write |           4 |   523.8 us |  9.34 us |  5.56 us |  7.8125 |  40.09 KB |
|   Full |   Read |           4 |   562.0 us | 10.74 us |  6.39 us | 21.4844 |  98.68 KB |
|   Full | Random |           4 |   584.6 us | 12.79 us |  8.46 us | 14.6484 |  69.38 KB |
|   Full |  Write |           1 | 2,084.2 us | 24.97 us | 16.52 us | 31.2500 | 156.76 KB |
|   Full |   Read |           1 | 2,135.2 us | 38.03 us | 25.16 us | 82.0313 | 391.14 KB |
|   Full | Random |           1 | 2,185.0 us | 27.74 us | 18.35 us | 58.5938 | 273.95 KB |